src/apps/server/pyproject.toml
[project]
name = "server"
version = "0.0.1"
readme = "README.md"
authors = [{ name = "arazzz" }]
requires-python = ">=3.11"
dependencies = [
	"pylsl==1.16.2",
	"bci-essentials==0.1.3",
	"fastapi",
	"python-dotenv",
	"python-socketio",
	"pyyaml",
	"rich",
	"numpy",
]
